Merge all changes from branch 13.x
Including PTB upadtes, fixes and addition of new modules

+++ SaitamaRobot/modules/sql/feds_sql.py
@@ -3,7 +3,7 @@ import ast
 
 from SaitamaRobot import dispatcher
 from SaitamaRobot.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
+from sqlalchemy import Boolean, Column, Integer, BigInteger, String, UnicodeText
 from telegram.error import BadRequest, Unauthorized
 
 
@@ -59,7 +59,7 @@ class BansF(BASE):
 
 class FedsUserSettings(BASE):
     __tablename__ = "feds_settings"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     should_report = Column(Boolean, default=True)
 
     def __init__(self, user_id):
